@Jaaslet 2023-04-16
I will close this PR. It's a big change, and we are already soon getting the advanced catch filters, which already will make you spend less time on ball animations