@xslk 2023-04-16
What about upping the speed of only the special balls that are at 1.25s catch time to make them slightly more appealing?